+++ src/main/ngapp/src/app/dataservices/virtualization/view-editor/view-canvas/view-canvas.component.html
@@ -1,5 +1,10 @@
 <div id="view-editor-canvas-editor">
 
+  <div class="view-editor-area"
+       id="views-list-container">
+    <app-views-list></app-views-list>
+  </div>
+
   <!-- ============================ -->
   <!-- ========== Canvas ========== -->
   <!-- ============================ -->
